requirements_v2.txt
# VoiceSnap v2 - Desktop Application Requirements

# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.24.0

# Audio
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13

# UI Framework
customtkinter>=5.2.0
pillow>=10.0.0

# System Integration
pynput>=1.7.6
pyperclip>=1.8.2
pystray>=0.19.5

# Database
python-dotenv>=1.0.0

# Platform-specific
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'
pywin32>=306; sys_platform == 'win32'
